Rename dta7400_week1_notebook.ipynb to Mack_week1.ipynb
A powerful and necessary part of modern software development is using Git and GitHub to keep track of changes and work on projects together. When I think back on the process, I can remember both the good and bad things that happened. Git is a good tool for keeping track of different versions of a project. It is very helpful to be able to keep track of changes, go back to earlier versions, and create branches for adding features or fixing bugs. It makes it easier for people to work together and keeps a good record of the project's history. Dealing with merge conflicts is a common problem that happens when more than one person changes the same file. Talking to team members about ongoing work on a regular basis and getting the most recent changes before making changes helps keep disagreements to a minimum. When conflicts do happen, it is important to use Git is tools to solve them in a planned way and let the team know about any changes.